feat(monitor): 監看模式改為靜態 12 欄網格佈局
header 以下的三段式手寫 flex（TimelineDock 全寬 → News Feed 40% +
IndicatorPanel 兩欄 → 底部三卡）改成單一可捲動的 12 欄 CSS grid，
14 個 widget 依排版沙盒定稿座標擺位。

- monitorLayout.ts：沙盒匯出的座標 SSOT（相容 react-grid-layout layout
  陣列）+ MONITOR_HIDDEN 過濾，版面調整回沙盒改、不動 JSX
- MonitorPanel：grid 容器（repeat(12, minmax(0,1fr)) / gridAutoRows 40px /
  gap 10px）+ widget id → 節點對照表，由 layout 陣列 map 出 cell
- 抽出 NewsFeedPanel（原 MonitorPanel 內嵌的 News Feed 欄）
- 抽出 HotspotsWidget / HourlyHistogramWidget / TriageWidget
  （照未合併分支 46218e5 原樣搬移，計算邏輯隨元件走）
- 刪除 IndicatorPanel：widget 全部上網格後不再有人引用
- cell 用 overflow:auto + 子節點 flex:1 0 auto：矮於格子就撐滿、
  高於格子就格內捲動，不裁切也不壓到下一列（LiveWall 是 16:9 磚，
  實高隨欄寬變動，固定 row span 無法通吃）

Header、面板高度拖拉、wall mode 定位、所有資料 fetch / RPC / 輪詢頻率
一律未動；PR #89 的 sparkline 與機場卡修正未觸碰。
