+++ docs/features/monitor-grid-static/backlog.md
@@ -1,0 +1,19 @@
+# Backlog — <feature-name>
+
+> 本 feature 的待辦。與全站 `.claude/memory/BACKLOG.md` 對應項編號要一致（例如 RE-1、RE-2）。
+
+## 進行中
+
+- [ ] **RE-X**：<描述> — <為什麼、下一步>
+
+## 待辦
+
+- [ ] **RE-Y**：<描述>
+
+## 已完成（近期）
+
+- [x] **RE-Z**：<描述> — PR #NN, YYYY-MM-DD
+
+## 已放棄 / 延後
+
+- **RE-W**：<描述> — 原因：...
